pyproject.toml
[project]
name = "mememind-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alabaster==1.0.0",
    "alembic>=1.17.1",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "apscheduler==3.11.0",
    "asgiref>=3.10.0",
    "attrs==25.3.0",
    "authlib==1.6.4",
    "babel==2.17.0",
    "bcrypt==5.0.0",
    "boto3>=1.40.63",
    "botocore>=1.40.63",
    "celery>=5.5.3",
    "certifi==2025.8.3",
    "cffi==2.0.0",
    "charset-normalizer==3.4.3",
    "chromadb>=1.3.3",
    "click==8.3.0",
    "cryptography==46.0.1",
    "cyclopts==3.24.0",
    "distro==1.9.0",
    "dnspython==2.8.0",
    "docstring-parser==0.17.0",
    "docutils==0.21.2",
    "ecdsa==0.19.1",
    "email-validator==2.3.0",
    "exceptiongroup==1.3.0",
    "fastapi[standard]==0.118.0",
    "fastmcp==2.12.4",
    "gradio>=5.49.1",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "idna==3.10",
    "imagesize==1.4.1",
    "iniconfig==2.1.0",
    "isodate==0.7.2",
    "jinja2==3.1.6",
    "jiter==0.11.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.1",
    "jsonschema-path==0.3.4",
    "jsonschema-specifications==2025.9.1",
    "langchain==1.0.2",
    "langchain-core==1.0.1",
    "langchain-openai==1.0.1",
    "langchain-text-splitters>=1.0.0",
    "langgraph==1.0.1",
    "langgraph-checkpoint==3.0.0",
    "langgraph-prebuilt==1.0.1",
    "langgraph-sdk==0.2.9",
    "langsmith==0.4.38",
    "lazy-object-proxy==1.12.0",
    "llama-cpp-python>=0.3.16",
    "loguru==0.7.3",
    "lxml>=6.0.2",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mcp==1.15.0",
    "mdurl==0.1.2",
    "more-itertools==10.8.0",
    "openai==2.6.1",
    "openapi-core==0.19.5",
    "openapi-pydantic==0.5.1",
    "openapi-schema-validator==0.6.3",
    "openapi-spec-validator==0.7.2",
    "orjson==3.11.4",
    "ormsgpack==1.11.0",
    "packaging==25.0",
    "pandas>=2.3.3",
    "parse==1.20.2",
    "passlib==1.7.4",
    "pathable==0.4.4",
    "playwright==1.55.0",
    "pluggy==1.6.0",
    "pyasn1==0.6.1",
    "pycparser==2.23",
    "pydantic==2.11.9",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.11.0",
    "pyee==13.0.0",
    "pygments==2.19.2",
    "pymysql==1.1.2",
    "pyperclip==1.11.0",
    "pytest==8.4.2",
    "pytest-asyncio==1.2.0",
    "python-docx>=1.2.0",
    "python-dotenv==1.1.1",
    "python-jose==3.5.0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.3",
    "redis>=7.0.1",
    "referencing==0.36.2",
    "regex==2025.10.23",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rfc3339-validator==0.1.4",
    "rich==14.1.0",
    "rich-rst==1.3.1",
    "roman-numerals-py==3.1.0",
    "rpds-py==0.27.1",
    "rsa==4.9.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "snowballstemmer==3.0.1",
    "sphinx==8.2.3",
    "sphinx-rtd-theme==3.0.2",
    "sphinxcontrib-applehelp==2.0.0",
    "sphinxcontrib-devhelp==2.0.0",
    "sphinxcontrib-htmlhelp==2.1.0",
    "sphinxcontrib-jquery==4.1",
    "sphinxcontrib-jsmath==1.0.1",
    "sphinxcontrib-qthelp==2.0.0",
    "sphinxcontrib-serializinghtml==2.0.0",
    "sqlalchemy==2.0.43",
    "sqlalchemy2-stubs==0.0.2a38",
    "asyncpg>=0.28.0",
    "sse-starlette==3.0.2",
    "starlette==0.48.0",
    "tenacity==9.1.2",
    "tiktoken==0.12.0",
    "torch>=2.2.2",
    "tqdm==4.67.1",
    "transformers>=4.57.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "tzlocal==5.3.1",
    "unstructured[docx,md,pdf,pdfminer,pptx,xlsx]>=0.18.18",
    "urllib3==2.5.0",
    "uv==0.8.22",
    "uvicorn==0.37.0",
    "werkzeug==3.1.1",
    "xxhash==3.6.0",
    "zstandard==0.25.0",
    "accelerate>=1.11.0",
    "pdfplumber>=0.11.8",
    "pymupdf>=1.26.6",
]
[tool.uv]
required-environments = [
    "sys_platform == 'darwin' and platform_machine == 'x86_64'"
]
